fix(homie-adapter.ts): ignore property meta tags for now
I noticed during testing, that these tags were added as properties to the thing.
This exposed some nasty internal information about the Thing itself that we probably
don't want to expose. Since we are not verifying any of the configured properties nor
setting a name, we can safely ignore them as for now!